PutTracking Line
The performance it not the same, but actually logical.. How can we make
it the same? TGoal heading at 100 meters on the path? only to the goal
afterwards? or chnage the Fact? The point of closest distance NEED to be
in front! - however, this is not important, as just going back to the
heaidng will also work!