Confused by the nature of turning
The TGoVel can be devired, for global, and body -> need to confirm The
angles can be derived as well. but. but. The turning? should also be on
the body axis. NOT the global axis. This turning is tehn shoukld be
cangeh into Global. Or at least, the V change should be transform to
global. The V change! in a small time value! not he the angle! well the
roll is stilll the angle.