Problem on recording the VOp point
Sometimes the point is no a point --> make the point clculation error.
what to do with the unvalid point? incl if there are actually no point?
seems its still a long way to go... :( But maybe the key is the way to
record them.. so the polygon can have a different number of point. and
what to do if zero point..