the cube8 is not simmtric
it is hard to find why. But the points resolution might be the reason.
so finding the same point maybe necessary. If the vel it kept, maybe it
will be symmetric? but then we need to change the logic. Why we didnt
use the initial method of just use the angle, keeping the magnitude the
same?